pyproject.toml
[tool.poetry]
name = "twitter-monitor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "twitter_monitor"}]

[tool.poetry.dependencies]
python = "^3.10"
tweepy = {extras = ["async"], version = "^4.12.1"}
python-dotenv = "^0.21.0"
nextcord = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
